refactor: 패키지 정리 및 MessageGenerator 추가
